Create jira-ticket-checker probot
This Github bot will enforce that any opened/edited pull requests are
linked to at least one Jira ticket. The ticket number can either be in
the pull request title or the branch name itself. Note: this bot
requires that a `.github/config.yml` file exists on the default branch
of either the target repository (with a `jiraBaseUrl` value) or in the
organization's `.github` repository.