Exclude Non-Feature Branches in Jira Ticket Check
`headBranchExcluded` is a boolean value based on the existance of a
regular expression string in the `.Github` `config.yml` file. The
regular expression determines what branches to include in the check.
Anything matching `/(develop$|master$|hotfix/|release/)/i` will be
excluded—so basically only feature branches will be required to
contain a Jira ticket number. `headBranchExcluded` defaults to false.

This PR is connected to this PR in the `.Github` repository:
https://github.com/artnetworldwide/.github/pull/2/files